+++ packages/library/src/commonMain/kotlin/io/realm/Realm.kt
@@ -64,8 +64,8 @@ class Realm private constructor(configuration: RealmConfiguration, dbPointer: Na
         this(configuration, RealmInterop.realm_open(configuration.nativeConfig))
 
     /**
-     * Modify the underlying Realm file in a suspendable transaction on the default Realm
-     * dispatcher.
+     * Modify the underlying Realm file in a suspendable transaction on the default Realm Write
+     * Dispatcher.
      *
      * The write transaction always represent the latest version of data in the Realm file, even if
      * the calling Realm not yet represent this.
@@ -78,7 +78,6 @@ class Realm private constructor(configuration: RealmConfiguration, dbPointer: Na
      * frozen before returning it.
      * @see [RealmConfiguration.writeDispatcher]
      */
-    // TODO Would we be able to offer a per write error handler by adding a CoroutineExceptinoHandler
     suspend fun <R> write(block: MutableRealm.() -> R): R {
         @Suppress("TooGenericExceptionCaught") // FIXME https://github.com/realm/realm-kotlin/issues/70
         try {
@@ -118,7 +117,7 @@ class Realm private constructor(configuration: RealmConfiguration, dbPointer: Na
 
     private suspend fun updateRealmPointer(newRealm: NativePointer, newVersion: VersionId) {
         realmPointerMutex.withLock {
-            log.debug("$version -> $newVersion")
+            log.debug("Updating Realm version: $version -> $newVersion")
             if (newVersion >= version) {
                 dbPointer = newRealm
                 version = newVersion
@@ -129,11 +128,20 @@ class Realm private constructor(configuration: RealmConfiguration, dbPointer: Na
     /**
      * Close this Realm and all underlying resources. Accessing any methods or Realm Objects after this
      * method has been called will then an [IllegalStateException].
+     *
+     * This will block until underlying Realms (writer and notifier) are closed, including rolling
+     * back any ongoing transactions when [close] is called. Calling this from the Realm Write
+     * Dispatcher while inside a transaction block will throw, while calling this by some means of
+     * a blocking operation on another thread (e.g. `runBlocking(Dispatcher.Default)`) inside a
+     * transaction cause a deadlock.
+     *
+     * @throws IllegalStateException if called from the Realm Write Dispatcher while inside a
+     * transaction block.
      */
     public override fun close() {
         // TODO Reconsider this constraint. We have the primitives to check is we are on the
         //  writer thread and just close the realm in writer.close()
-        writer.checkInTransaction("Cannot close in a transaction block")
+        writer.checkInTransaction("Cannot close the Realm while inside a transaction block")
         writer.close()
         super.close()
         // TODO There is currently nothing that tears down the dispatcher