+++ test/src/androidTest/kotlin/io/realm/shared/RealmTests.kt
@@ -207,7 +207,7 @@ class RealmTests {
 
     @Test
     @Suppress("invisible_member")
-    fun simultaneousWritesAreSerialized() = runBlocking {
+    fun simultaneousWritesAreAllExecuted() = runBlocking {
         val jobs: List<Job> = IntRange(0, 9).map {
             launch {
                 realm.write {
@@ -227,18 +227,27 @@ class RealmTests {
     @Test
     @Suppress("invisible_member")
     fun writeBlockingWhileWritingIsSerialized() = runBlocking {
-        val mutex = Mutex(true)
-        val exit = Mutex(true)
-        val write = async {
+        val writeStarted = Mutex(true)
+        val writeEnding = Mutex(true)
+        val writeBlockingQueued = Mutex(true)
+        async {
             realm.write {
-                mutex.unlock()
-                PlatformUtils.sleep(1.milliseconds)
-                exit.unlock()
+                writeStarted.unlock()
+                while(writeBlockingQueued.isLocked) {
+                    PlatformUtils.sleep(1.milliseconds)
+                }
+                writeEnding.unlock()
             }
         }
-        mutex.lock()
-        realm.writeBlocking {
-            assertFalse { exit.isLocked }
+        writeStarted.lock()
+        runBlocking {
+            val async = async {
+                realm.writeBlocking {
+                    assertFalse { writeEnding.isLocked }
+                }
+            }
+            writeBlockingQueued.unlock()
+            async.await()
         }
     }
 
@@ -257,22 +266,21 @@ class RealmTests {
 
     @Test
     @Suppress("invisible_member")
-    // Flaky:
     fun closeCausesOngoingWriteToThrow() = runBlocking {
-        val mutex = Mutex(true)
-        val exit = Mutex(true)
+        val writeStarted = Mutex(true)
         val write = async {
             assertFailsWith<RuntimeException> {
                 realm.write {
-                    mutex.unlock()
+                    writeStarted.unlock()
                     copyToRealm(Parent())
+                    // realm.close is blocking until write block is done, so we cannot wait on
+                    // specific external events, so just sleep a bit :/
                     PlatformUtils.sleep(Duration.Companion.milliseconds(100))
                 }
             }
         }
-        mutex.lock()
+        writeStarted.lock()
         realm.close()
-        exit.unlock()
         assert(write.await() is RuntimeException)
         realm = Realm.open(configuration)
         assertEquals(0, realm.objects<Parent>().size)